+++ sdks/csharp/src/EventHandling/Backend.cs
@@ -1,0 +1,24 @@
+using System;
+
+namespace SpacetimeDB.EventHandling
+{
+    public static class Backend
+    {
+        internal static bool UseNativeDispatch { get; private set; } = true;
+        private static IEventListenersFactory? CustomFactory { get; set; }
+
+        public static void UseNativeEvents()
+        {
+            UseNativeDispatch = true;
+            CustomFactory = null;
+        }
+
+        public static void UseCustomListeners(IEventListenersFactory? factory = null)
+        {
+            UseNativeDispatch = false;
+            CustomFactory = null;
+        }
+
+        internal static IEventListeners<T> Create<T>() where T : Delegate => CustomFactory?.Create<T>() ?? new EventListeners<T>();
+    }
+}
